Tests/LibWeb/Text/expected/WebAnimations/transitions/transition-on-pseudo-element.txt

running animations count (0) should be zero
